Minor fixes to balloons table, webpage, cmdline tool: consistently use tinyint for booleans, lowercase table headers, remove TD tags to fix HTML, only query balloons not done yet, sort by submitid.
git-svn-id: https://secure.a-eskwadraat.nl/svn/domjudge/trunk@3746 caf94f87-7bd6-0310-88dc-dee8202e2269

+++ sql/mysql_db_structure.sql
@@ -14,7 +14,7 @@
 CREATE TABLE `balloon` (
   `balloonid` int(4) unsigned NOT NULL AUTO_INCREMENT COMMENT 'Unique ID',
   `submitid` int(4) unsigned NOT NULL COMMENT 'Submission for which balloon was earned',
-  `done` int(1) unsigned NOT NULL DEFAULT '0' COMMENT 'Has been handed out yet?',
+  `done` tinyint(1) unsigned NOT NULL DEFAULT '0' COMMENT 'Has been handed out yet?',
   PRIMARY KEY (`balloonid`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COMMENT='Balloons to be handed out';
